Set cookies on login and include expiration in session query

+++ api/databases/db_ingredient.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Date, DateTime
+from sqlalchemy import Column, Integer, String, Date, DateTime, Float
 from sqlalchemy.ext.declarative import declarative_base
 
 Base = declarative_base()
@@ -8,7 +8,7 @@ class DBIngredient(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     name = Column(String, index=True)
-    amount = Column(Integer)
+    amount = Column(Float)
     unit = Column(String)
     expiry_date = Column(Date)
     storage = Column(String)
